+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support multiple task functions per worker.
 
+### Changed
+
+- Use O(1) queue implementation for tasks queueing.
+
 ## [2.4.11] - 2023-04-23
 
 ### Changed